Refactor XTCE conversion
