+++ crates/hermes-data/Cargo.toml
@@ -10,5 +10,5 @@ thiserror = "2.0"
 tracing = "0.1.44"
 
 [dev-dependencies]
-test-log = {version = "0.2", default-features = false, features = ["trace"]}
+test-log = { version = "0.2", default-features = false, features = ["trace"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
