+++ crates/hermes-data/tests/common/mod.rs
@@ -1,5 +1,5 @@
-use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::Arc;
 use tracing::Level;
 use tracing_subscriber::layer::SubscriberExt;
